Decoupled app structure
