+++ index.html
@@ -56,13 +56,7 @@
         </section>
 
         <section class="authentication" id="auth_modal" data-show="false">
-
-            
-
-            <div class="authentication__content" id="auth_modal--content">
-                <span id="closeModal">+</span>
-            </div>
-
+            <div class="authentication__content" id="auth_modal--content"></div>
         </section>
 
     </main>
