+++ lib/index.js
@@ -1,23 +1,20 @@
 import '../scss/style.scss';
+import { listenForAuthSwitch, listenForSignin, 
+        listenForSignOut, listenForSignup, 
+        listenForTodosConnectionStatus, listenForTodosConnectionFailed
+ } from './listeners';
 import { 
-    authenticate, clearCompletedTodo, 
-    createTaskSocket, createTodo,
+    clearCompletedTodo, createTodo,
     deleteTodo, updateTodo,
     registerTaskConnectionListener, registerTasksUpdateListener,
-    registerTaskFailedConnectionListener, USER_TOKEN } from './script';
+    registerTaskFailedConnectionListener, USER_TOKEN,
+    handleSignOut } from './script';
 
 import { alreadySignedInTemplate, signUpTemplate, signInTemplate,renderTodosTemplate, initiateCloseModal, taskTemplate } from './templates';
 
 let filter = 'all';
 let todos = [];
 
-// Keep a watch over all todos, and responed to any event on them
-// Features: delete todo and update todo
-const listenForTasksUpdate = (payload)=>{
-    // payload is the tasks
-    todos = payload;
-    renderTodos();
-};
 
 function trackTodos() {
     const cancels = document.querySelectorAll('[data-cancel_id]');
@@ -71,10 +68,11 @@ const renderTodos = () => {
 
 }
 
-const renderAuthModal = (type=null, close=false) =>{    
+const renderAuthModal = (type=null, toClose=false) =>{    
     const modal = document.getElementById("auth_modal");
     const modal_content = document.getElementById("auth_modal--content");
-    if(close){
+    
+    if(toClose){
         modal.setAttribute("data-show", "false");
         return
     }
@@ -87,7 +85,7 @@ const renderAuthModal = (type=null, close=false) =>{
         modal_content.innerHTML = alreadySignedInTemplate(AUTH_DATA.username);
         modal.setAttribute("data-show", "true");
         initiateCloseModal(modal); // a working close modal
-        listenForSignOut();
+        listenForSignOut(signOut);
         return
     }
     
@@ -105,141 +103,23 @@ const renderAuthModal = (type=null, close=false) =>{
     listenForAuthSwitch();
 }
 
-function listenForAuthSwitch (){
-
-    document.getElementById("switchContent").addEventListener('click',(e)=>{
-        e.preventDefault();
-        // console.log(e.target.dataset);
-        const modal_content = document.getElementById("auth_modal--content");
-        const navTo = e.target.dataset.nav;        
-        // Switch content
-        if (navTo === 'signUp') {
-            modal_content.innerHTML = signUpTemplate;
-            listenForSignup();
-        }
-        else if (navTo === 'signIn') {
-            modal_content.innerHTML = signInTemplate;
-            listenForSignin();
-        };
-
-        listenForAuthSwitch();
-    });
-}
-
-function listenForSignin(){
-    const form = document.getElementById("authSignIn");
-
-    form.onsubmit = (e) =>{
-        e.preventDefault();
-        const form_data = e.target;
-
-        const username = form_data.usernameEmail.value;
-        const password = form_data.password.value;
-
-        const userData = {
-            username,
-            password
-        };
-        
-        handleAuth(userData, 'signin', form_data);
-    }
-}
-
-function listenForSignup(){
-    const form = document.getElementById("authSignUp");
-
-    form.onsubmit = (e) =>{
-        e.preventDefault();
-        const form_data = e.target;
-
-        const username = form_data.username.value;
-        const email = form_data.email.value;
-        const password = form_data.password.value;
-
-        const userData = {
-            username,
-            email,
-            password
-        };
-
-        handleAuth(userData, 'signup', form_data);
-    }
-
-}
+// Keep a watch over all todos, and responed to any event on them
+// Features: delete todo and update todo
+const listenForTasksUpdate = (payload)=>{
+    // payload is the tasks
+    todos = payload;
+    renderTodos();
+};
 
 
-const handleAuth = (userData, mode, form_data)=>{
-    form_data.setAttribute('data-loading', 'true');
-    form_data.formButton.innerText = mode === 'signin' ? "Signing In..." : "Signing Up...";
-    form_data.formButton.setAttribute('type','button');
-
-    document.getElementById("form_err").innerText = "";
-    
-    authenticate(userData, mode, handleAuthAcknowledgement);
-}
-
-const handleAuthAcknowledgement = (payload)=>{
-    const {data, error, mode} = payload;
-
-    if (error){
-        console.error("Error occured:", error);
-        // renderTodosError(error);
-        document.getElementById("form_err").innerText = error;
-        
-        const form_data = document.getElementById(mode === 'signin' ? "authSignIn" : "authSignUp");
-        form_data.setAttribute('data-loading', 'false');
-        form_data.formButton.innerText = mode === 'signin' ? "Sign In" : "Sign Up";
-        form_data.formButton.setAttribute('type','submit');
-        return;
-    }
-
-
-    if(mode === 'signin'){
-        document.getElementById("auth").innerText = data.username;
-
-        localStorage.setItem(USER_TOKEN, JSON.stringify(data));
-
-        renderAuthModal(null, true); // close Modal
-
-        const {token} = data;
-
-        createTaskSocket(token);
-    }
-    else { // signup
-        document.getElementById("switchContent").click();
-        document.querySelector('input[name="usernameEmail"]').value = data.username;
-    }
-
-    
-
-}
-
-
-const listenForTodosConnectionStatus = (connecting=true)=>{
-    // connected: indicate if trying to reach server (short)
-    if (connecting) document.querySelector(".task-items").setAttribute('data-loading', 'true');
-    else document.querySelector(".task-items").setAttribute('data-loading', 'false');
-}
-
-const listenForTodosConnectionFailed = ()=>{
-    return signOut();
-}
-
-
-
-const signOut = ()=>{
-    localStorage.removeItem(USER_TOKEN);
+const signOut = ()=> handleSignOut(()=>{
+    // localStorage.removeItem(USER_TOKEN);
     document.getElementById("auth").innerText = "Login";
     todos = []
 
     renderTodos();
     renderAuthModal();
-    createTaskSocket(null);
-}
-
-function listenForSignOut(){
-    document.getElementById("signOut").addEventListener('click',signOut);
-}
+});
 
 // Update by filter
 const updateByFilter = () => {
@@ -267,44 +147,6 @@ const updateByFilter = () => {
     renderTodos();
 }
 
-function ensureAuthentication(){
-    // Check for auth
-    // if user data exist, connect and fetch data
-    // otherwise authenticate
-    let authdata = null;
-
-    try{
-        authdata = JSON.parse(localStorage.getItem(USER_TOKEN));
-    }catch{ // cannot resolve cached auth data
-        // remove whatever is there
-        localStorage.removeItem(USER_TOKEN);
-        renderAuthModal(); // render auth
-        return;
-    }
-
-    // if not authdata, render auth
-    if(!authdata){
-        renderAuthModal(); // render auth
-        return;
-    }
-
-    // if there is authdata
-    const {token} = authdata; // get user token
-
-    if (!token ) { // if user token is absent
-        localStorage.removeItem(USER_TOKEN); // remove auth data
-        document.getElementById("auth").innerText = "Login";
-        renderAuthModal(); // render auth
-        return
-    }
-
-    // at this point, all is set from cache
-    document.getElementById("auth").innerText = authdata.username;
-    
-    createTaskSocket(token); // try reconnect
-}
-
-
 // Run when the browser is done loading
 window.onload = (e) => {
     // Run function when the browser loads
@@ -317,11 +159,9 @@ window.onload = (e) => {
         renderAuthModal();
     });
 
-    ensureAuthentication();
-
     registerTaskConnectionListener(listenForTodosConnectionStatus);
     registerTasksUpdateListener(listenForTasksUpdate);
-    registerTaskFailedConnectionListener(listenForTodosConnectionFailed);
+    registerTaskFailedConnectionListener((error)=> listenForTodosConnectionFailed(error, signOut));
 
 }
