+++ lib/script.js
@@ -6,11 +6,14 @@ if(!url){
     console.error("No Url specified");
 }
 
-const authSocket = io(`${url}/auth`);
-const tasksSocket = io(`${url}/task`, {query:{ token: "sample"}});
 
 export const USER_TOKEN = 'cpc-todi-token'; // locator on local storage
 const PLATFORM = 'web';
+const {token:userAuthToken} = JSON.parse(localStorage.getItem(USER_TOKEN)) || {};
+
+const authSocket = io(`${url}/auth`);
+const tasksSocket = io(`${url}/task`, {query:{ token: userAuthToken}});
+
 
 const EVENTS = {
     connection:'connect',
@@ -47,6 +50,11 @@ const EVENTS = {
 //     console.log("Auth socket Disonnected");
 // })
 
+// authSocket.on(EVENTS.error, (err)=>{
+//     console.log("Auth socket Error", err);
+// })
+
+
 // Task connection
 tasksSocket.on(EVENTS.connection, ()=>{
     // console.log("Task socket Connected");
@@ -60,8 +68,8 @@ tasksSocket.on(EVENTS.connection, ()=>{
 export const registerTaskFailedConnectionListener = (cb)=>{
     // when just unable to connect
     tasksSocket.on(EVENTS.error, (err)=>{
-        // console.log("Task socket error", err)
-        cb()
+        console.log("Task socket error", err)
+        cb(err);
     });
 }
 
@@ -88,6 +96,18 @@ export const registerTasksUpdateListener = (cb)=>{
     });
 }
 
+export const handleSignOut = (cb)=>{
+    // Reconnect auth socket
+    authSocket.connect();
+    // disconnect task socket
+    tasksSocket.disconnect();
+    // in callback, delete token, and reset username and todos
+
+    localStorage.removeItem(USER_TOKEN);
+
+    cb()
+}
+
 
 
 export const authenticate = (data, mode, cb)=>{
@@ -110,15 +130,16 @@ export const authenticate = (data, mode, cb)=>{
         authSocket.emit(EVENTS.signin_account, data, (payload)=> {
             cb({...payload, mode});
             // Disconnect
-            // authSocket.disconnect();
+            authSocket.disconnect();
         });
     }
 }
 
 // Reconnect maybe after signin
-export const createTaskSocket = (token) =>{
+export const createTaskSocket = () =>{
 
-    if(!token) return tasksSocket.disconnect();
+    // if(!token) return tasksSocket.disconnect();
+    const {token} = JSON.parse(localStorage.getItem(USER_TOKEN)) || {};
 
     tasksSocket.query.token = token;
     // console.log(tasksSocket);
